refactor: merge collection and databasing
