--- src/databasing/config.py
@@ -1,11 +1,0 @@
-import os
-from dotenv import load_dotenv
-
-load_dotenv()
-
-DB_CONFIG = {
-    "host": os.getenv("DB_HOST","localhost"),
-    "user": os.getenv("DB_USER", "root"),
-    "password": os.getenv("DB_PASSWORD", ""),
-    "database": os.getenv("DB_NAME", "stock_db")
-}
