+++ temp_create_stockdb.sql
@@ -1,0 +1,18 @@
+CREATE DATABASE IF NOT EXISTS stocks_db DEFAULT CHARSET utf8mb4;
+USE stocks_db;
+
+CREATE TABLE IF NOT EXISTS stock_prices (
+  id BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
+  ticker VARCHAR(16) NOT NULL,
+  dt DATETIME NOT NULL,
+  open DOUBLE NULL,
+  high DOUBLE NULL,
+  low DOUBLE NULL,
+  close DOUBLE NULL,
+  adj_close DOUBLE NULL,
+  volume BIGINT NULL,
+  `interval` VARCHAR(8) NOT NULL DEFAULT '1d',
+  ingested_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
+  PRIMARY KEY (id),
+  UNIQUE KEY ux_ticker_dt_interval (ticker, dt, `interval`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
